split `sandbox` requests into `config` and `job`
